Concertando bugs por conta da alteração de long para int e comentando todo o crud de contato

+++ src/main/java/com/example/demo/controllers/CuidadorController.java
@@ -36,7 +36,7 @@ public class CuidadorController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<CuidadorDTO> buscarPorId(@PathVariable Long id) {
+    public ResponseEntity<CuidadorDTO> buscarPorId(@PathVariable Integer id) {
         return ResponseEntity.ok(service.buscarPorId(id));
     }
 
@@ -47,12 +47,12 @@ public class CuidadorController {
     }
 
     @PutMapping("/atualizar/{id}")
-    public ResponseEntity<CuidadorDTO> atualizar(@PathVariable Long id, @RequestBody CuidadorDTO dto) {
+    public ResponseEntity<CuidadorDTO> atualizar(@PathVariable Integer id, @RequestBody CuidadorDTO dto) {
         return ResponseEntity.ok(service.atualizar(id, dto));
     }
 
     @DeleteMapping("/deletar/{id}")
-    public ResponseEntity<Void> deletar(@PathVariable Long id) {
+    public ResponseEntity<Void> deletar(@PathVariable Integer id) {
         service.inativar(id);
         return ResponseEntity.noContent().build();
     }
